made parts easier to read, also some bot stuff
Re-wrote some parts of the BlackJackGame class to be easier to read.
Also ,players' names should be saved as their usernames as an @mention. Don't know if that worked or not, but did it anyway. The /blackjack also now starts a game with just the player who did the command. Also tried to add some joining a game functionality.
I also Commented out a large section of the code, so the game itself will not run currently unless you uncomment it.