Added getHandStrength()
added a method called getHandStrength() along with 10 other methods to calculate the strength of the best 5 card poker hand that can be made from a list of cards given to the method.
IT IS NOT FINISHED. The straight, straight flush, and royal flush methods are empty and need to be written. The rest of the code should be fine though.
there is a comment block above the section of code that details how the code stores the strength of the hand and how it should be returned.